поменял double на BigDecimal для модели обменного курса

+++ src/main/java/com/github/petrovyegor/currencyexchange/model/ExchangeRate.java
@@ -1,18 +1,20 @@
 package com.github.petrovyegor.currencyexchange.model;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
     private int id;
     private int baseCurrencyId;
     private int targetCurrencyId;
-    private double rate;
+    private BigDecimal rate;
 
-    public ExchangeRate(int baseCurrencyId, int targetCurrencyId, double rate) {
+    public ExchangeRate(int baseCurrencyId, int targetCurrencyId, BigDecimal rate) {
         this.baseCurrencyId = baseCurrencyId;
         this.targetCurrencyId = targetCurrencyId;
         this.rate = rate;
     }
 
-    public ExchangeRate(int id, int baseCurrencyId, int targetCurrencyId, double rate) {
+    public ExchangeRate(int id, int baseCurrencyId, int targetCurrencyId, BigDecimal rate) {
         this.id = id;
         this.baseCurrencyId = baseCurrencyId;
         this.targetCurrencyId = targetCurrencyId;
@@ -43,11 +45,11 @@ public class ExchangeRate {
         this.targetCurrencyId = targetCurrencyId;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
 }
